requirements.txt
beautifulsoup4==4.11.1
django-ckeditor==6.2.0
django-compressor==4.4
django-debug-toolbar==3.2.4
gunicorn==20.1.0
ipython==7.31.0
pillow==9.0.0
pip-chill==1.0.1
psycopg2-binary==2.9.1
python-dotenv==0.20.0
python-memcached==1.59
pytz==2021.3
sentry-sdk==1.28.1
